Fix paths and the variable "iRun" in GenerateRun function.
Paths were defined for linux only, and are updated to 'os.path.join()' version.
The variable iRun is passed to the function GenerateRun,
and causes index out of range error in line 86.
I personally think passing "iRun" in this way is not an ideal way
as the variable is considered to be used as an iterator.
Fixed by, simply, setting index as -1 instead of "iRun".
I think it's the intended way of running,
according to the fact that length of the "_randStates" and "iRun" are the same
(10, when testing with the main function).
Besides, in the main function, the variable "ds" is used without definition.